+++ Entities/Models/ApplicationRole.cs
@@ -10,11 +10,18 @@ namespace Entities.Models
     public class ApplicationRole : IdentityRole<int>
     {
         /// <summary>
+        /// The number of matches for each day users with this subscription may see.
+        /// </summary>
+        public int DailyMatchesLimit { get; set; } = 0;
+
+        /// <summary>
         /// Navigational Property
         /// </summary>
         public virtual ICollection<PaddlePlanRole> PaddlePlanRoles { get; set; }
-
         public ApplicationRole() { }
-        public ApplicationRole(string roleName) : base(roleName) { }
+        public ApplicationRole(string roleName, int dailyMatchesLimit) : base(roleName)
+        {
+            DailyMatchesLimit = dailyMatchesLimit;
+        }
     }
 }