+++ MentorInterface/Controllers/MatchSelection/MatchSelectionController.cs
@@ -29,16 +29,22 @@ namespace MentorInterface.Controllers.MatchSelection
         /// User Manager
         /// </summary>
         private readonly UserManager<ApplicationUser> _userMananger;
+        private readonly RoleManager<ApplicationRole> _roleManager;
+        private readonly IRoleHelper _roleHelper;
 
         /// <summary>
         /// Create the controller and inject the HTTPClient factory.
         /// </summary>
         public MatchSelectionController(
             IHttpClientFactory clientFactory,
-            UserManager<ApplicationUser> userManager)
+            UserManager<ApplicationUser> userManager,
+            RoleManager<ApplicationRole> roleManager,
+            IRoleHelper roleHelper)
         {
             _clientFactory = clientFactory;
             _userMananger = userManager;
+            _roleManager = roleManager;
+            _roleHelper = roleHelper;
         }
 
         /// <summary>
@@ -49,24 +55,16 @@ namespace MentorInterface.Controllers.MatchSelection
         public async Task<IActionResult> MatchSelectionAsync(long steamId)
         {
             var user = await _userMananger.GetUserAsync(User);
-            var dailyLimit = await GetDailyLimitAsync(user);
+
+            var dailyMatchesLimit = await _roleHelper.GetDailyMatchesLimitAsync(user);
 
             var client = _clientFactory.CreateClient(ConnectedServices.MatchRetriever);
 
             HttpRequestMessage message = new HttpRequestMessage(
                 HttpMethod.Get,
-                $"v1/public/single/{steamId}/matchselection?dailyLimit={dailyLimit}");
+                $"v1/public/single/{steamId}/matchselection?dailyLimit={dailyMatchesLimit}");
 
             return await ForwardHttpRequest(client, message);
-        }
-
-        private async Task<int> GetDailyLimitAsync(ApplicationUser user)
-        {
-            var roles = await _userMananger.GetRolesAsync(user);
-            return 3;
-            // TODO: Return depending on role
-            //if(roles.Contains())
-            throw new NotImplementedException();
         }
     }
 }