+++ MentorInterface/Models/UserIdentity.cs
@@ -1,4 +1,5 @@
-﻿using MentorInterface.Helpers;
+﻿using Entities;
+using MentorInterface.Helpers;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -9,9 +10,8 @@ namespace MentorInterface.Models
     public class UserIdentity
     {
         public int ApplicationUserId { get; set; }
-
         public long SteamId { get; set; }
-
         public SubscriptionType SubscriptionType { get; set; }
+        public int DailyMatchesLimit { get; set; }
     }
 }