+++ MentorInterface/Paddle/PaddlePlanManager.cs
@@ -11,17 +11,27 @@ using System.Threading.Tasks;
 
 namespace MentorInterface.Paddle
 {
+    /// <summary>
+    /// Helper class for managing PaddlePlans.
+    /// </summary>
     public static class PaddlePlanManager
     {
         /// <summary>
-        /// List of all PaddlePlans and the roles they enable.
+        /// List of all PaddlePlans.
         /// </summary>
-        public static readonly List<PaddlePlanRoleBind> ProductionBinds = new List<PaddlePlanRoleBind>
+        public static readonly List<PaddlePlanRoleBind> ProductionPlans = new List<PaddlePlanRoleBind>
         {
-            new PaddlePlanRoleBind(583755, new List<string> {Subscriptions.Premium }),
-            new PaddlePlanRoleBind(583756, new List<string> {Subscriptions.Ultimate }),
+            // Premium
+            new PaddlePlanRoleBind(new PaddlePlan(586570, Entities.SubscriptionType.Premium, 1, 6.99), new List<string> { RoleCreator.Premium.Name}),
+            new PaddlePlanRoleBind(new PaddlePlan(586571, Entities.SubscriptionType.Premium, 3, 4.99), new List<string> { RoleCreator.Premium.Name}),
+            new PaddlePlanRoleBind(new PaddlePlan(586572, Entities.SubscriptionType.Premium, 6, 4.59), new List<string> { RoleCreator.Premium.Name}),
+
+            // Ultimate            
+            new PaddlePlanRoleBind(new PaddlePlan(586580, Entities.SubscriptionType.Ultimate, 1, 16.99), new List<string> { RoleCreator.Ultimate.Name}),
+            new PaddlePlanRoleBind(new PaddlePlan(586581, Entities.SubscriptionType.Ultimate, 3, 14.99), new List<string> { RoleCreator.Ultimate.Name}),
+            new PaddlePlanRoleBind(new PaddlePlan(586582, Entities.SubscriptionType.Ultimate, 6, 14.59), new List<string> { RoleCreator.Ultimate.Name}),
         };
-        
+
 
         /// <summary>
         /// Writes PaddlePlans and binds between PaddlePlans and ApplicationRoles to database for the Plans.
@@ -33,21 +43,21 @@ namespace MentorInterface.Paddle
             var applicationContext = serviceProvider.GetRequiredService<ApplicationContext>();
             var roleMananger = serviceProvider.GetRequiredService<RoleManager<ApplicationRole>>();
 
-            foreach (var roleBind in binds)
+            foreach (var bind in binds)
             {
-                if (!applicationContext.PaddlePlan.Any(x => x.PlanId == roleBind.PlanId))
+                var paddlePlan = bind.Plan;
+                if (!applicationContext.PaddlePlan.Any(x => x.PlanId == paddlePlan.PlanId))
                 {
                     // Create PaddlePlan
-                    var paddlePlan = new PaddlePlan { PlanId = roleBind.PlanId };
                     applicationContext.PaddlePlan.Add(paddlePlan);
 
                     // Create a PaddlePlanRole for each role this plan includes and write to database
-                    foreach (var roleName in roleBind.RoleNames)
+                    foreach (var roleName in bind.RoleNames)
                     {
                         var roleId = roleMananger.FindByNameAsync(roleName).Result.Id;
                         var paddlePlanRole = new PaddlePlanRole
                         {
-                            PlanId = roleBind.PlanId,
+                            PlanId = paddlePlan.PlanId,
                             RoleId = roleId,
                         };