+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+##[0.2.3] - ?
+### Added
+- Logic for computing DailyMatchesLimit
+- Controller for handling subscriptions
+
+## Changed
+- PaddlePlans and Roles are now created differently and coupled tighter than before
+- Cleared all migrations and start with fresh ones
+
+
 ## [0.2.2] - 2020-03-11
 ### Changed
 - Bugfix: Seed Database after migrations